requirements.txt
# astrapi-flask-ui V2 – Abhängigkeiten
# Installation: pip install -r requirements.txt

# ── Web-Framework ──────────────────────────────────────────────
flask>=3.0
jinja2>=3.1

# ── FastAPI + ASGI (für main.py / kombinierten Betrieb) ───────
fastapi>=0.115
uvicorn[standard]>=0.30
a2wsgi>=1.10            # WSGIMiddleware: Flask in FastAPI einbetten
pydantic>=2.0           # Datenmodelle in FastAPI-Routers

# ── Konfiguration ──────────────────────────────────────────────
pyyaml>=6.0             # Navigation + Schema-YAMLs
